refactor: 급여계산표 로직을 커스텀 훅으로 분리 #21
